[US_TN] Update safekeeping PFI normalization logic (Recidiviz/recidiviz-data#41869)
## Description of the change

Previously, IP normalization would stop overriding PFI with
TEMPORARY_CUSTODY as soon as it reached a period with admission reason
NEW_ADMISSION. This change adds another admission reason, REVOCATION, to
the set of admission reasons that mark the end of a safekeeping period.

This change also updates the way we check for these admission reasons to
correctly handle simultaneous periods. Previously, we would only check
the period being normalized to see if the admission reason marked the
end of a safekeeping period. Now, we'll look at all the periods with the
same admission date as the period being normalized, and if any of them
have an admission reason of NEW_ADMISSION or REVOCATION, then the
safekeeping period will end.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1d5f34fce65d64169784a5a37b49069fa1fb7170