[Email Reminders] Add testing functions (Recidiviz/recidiviz-data#41912)
## Description of the change

> Add testing functions for email reminders script. Added a total of
three tests: `getAuthMenuItem()`, `checkLoginStatusMenuItem()`, and
`testGetUserLoginInfo()`. `getAuthMenuItem()` is a menu button on the
Spreadsheet that retrieves the Auth0 Token, allowing for easy debugging.
`checkLoginStatusMenuItem()` is also a menu button that makes it more
intuitive for non-engineers to check users' login statuses.
`testGetUserLoginInfo()` is a testing function that checks if
`getUserLoginInfo` is returning the correct shape.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#41580

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 049806c97245e0973a999635f308b5f6f51c6265