Creates MA EGT views and hydrates person_projected_date_sessions (Recidiviz/recidiviz-data#42397)
## Description of the change

This PR creates the downstream views for MA EGT data to be used for the
JII tablet tool. The following are the views that are created as part of
this PR.

1. `analyst_data.us_ma_earned_credit_activity_preprocessed` - references
raw data and logs all of the credit activity in a month for each person.
Data is restructured so that all 3 types of credits (EGT, boosts, and
completion) are stored in the same column.
2. `analyst_data.earned_credit_activity` - a state agnostic version of
the above view that can be used to store data for other states as well.
The activity attributes (rating, activity name, etc) are stored in a
json since these are likely state specific.
3. `analyst_data. us_ma_person_projected_date_sessions_preprocessed` -
sessionizes and manipulates the the data in `analyst_data.
us_ma_projected_dates` so that can be unioned in with
`sentence_sessions.person_projected_date_sessions`. The main
manipulation is that the original and max release dates from US_MA are
both put in the same column `group_projected_full_term_release_date_max`
and a new row is created to log the initial value and track changes made
on this field just as we will as we receive updated data.

Following this PR, will be a PR that pulls the dates from
person_projected_date_sessions for a product view (will do a comparison
of the initial date to the current date to essentially create the
original and adjusted max release dates with logic that is state
agnostic).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6a7e44df7eba40a46f8373211e1cc4b6c595ce9b