[Testing] Combine View and Mapping tests (Recidiviz/recidiviz-data#42494)
## Description of the change

This PR folds the StateIngestMappingTestCase functionality into the
StateIngestViewTestCase and removes all mapping tests.

I updated the structure test to account for this change, which then
caught a few straggling v1 mappings that needed to be migrated.

When these were updated, it revealed that some parsing tests were
breaking from a known issue with the emulator (it doesn't quote dates
correctly in JSON). I added a workaround to quote JSON values and left a
TODO with that issue. (And to any IEs that read this, please feel free
to ping me with issues like this!!!)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f2195683e8af256401c9eeb5606759ad31cdf966