[Auth] Exclude facilities users from automatic blocking (Recidiviz/recidiviz-data#40565)
## Description of the change

Excludes users with any facilities role from automatic
blocking--assuming that any relevant role will have the word
"facilities" in it.

Question about the context behind this ticket: Is the roster sync only
used for supervision staff? So someone who used to have a supervision
role will be in `Roster` from past roster syncs, and then they move to a
facilities role and are no longer included in the roster sync but are
still a valid user.

If that's true, when the facilities role gets added, does it always or
at least sometimes add that role via a `UserOverride` record, so that we
need to also check the roles field in `UserOverride` for the facilities
role?

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#40441

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 47e32308b1f596b801518ef844719d35b955472c