add standard label to dependabot PRs (Recidiviz/recidiviz-data#42597)
## Description of the change

I noticed PRs like
https://github.com/Recidiviz/recidiviz-data/pull/42557 weren't being
automatically merged because they were failing the Enforce Required Pull
Request Labels check. This specifies the labels that it automatically
applies, and adds one of our standard labels to the mix so that it can
pass the check.

I haven't tested this, but it looks like it should work according to
https://docs.github.com/en/code-security/dependabot/working-with-dependabot/dependabot-options-reference#labels--.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 974b5ed2ee8895591ad85f9a603024f08d5ca0ed