[US_AR] Remove certain locations from allowed list of facility locations for building IPs (Recidiviz/recidiviz-data#42734)
## Description of the change

In the AR incarceration periods view, we define a list of locations
(based on the location type) that count as facilities where someone can
be incarcerated, because the raw table used to build incarceration
periods includes movements between locations that don't actually count
as facilities and we don't want to ingest incarceration periods for
these locations. This change narrows this list slightly to exclude
locations with type C1 (ACC Director's Office) or C9 (ACC Interstate
Compact), since these locations appear to only report movements for
people on supervision, not incarcerated people. These locations weren't
excluded originally because ACC jurisdiction _does_ include some people
in facilities such as community corrections centers, but locations with
C1/C9 location types don't report movements for these people.

Also adds a special exclusion for location `0141218`, which is named
'Unknown location'. This location needs to be specified using its ID,
since its location type (I4, 'Other State Law Enforc. Agency') _is_ a
valid facility type which can used for people incarcerated under other
states' jurisdiction. Location `0141218` is an exception, and only shows
up in movements where people are discharged to other states. As such, it
doesn't count as a location where someone can actually be incarcerated,
and therefore needs to be excluded.

Now, when someone moves from a facility to a location with location type
C1/C9 or location ID of `0141218`, there will still be a period for that
person ending in release to the C1/C9 or `0141218` location, but there
will no longer be a subsequent period ingested where the person starts
in the C1/C9 or `0141218` location.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e084cd5dd88944d8b0f6e61978fb6f603bf7f860