[US_AZ] Temporarily remove mental health assessment criteria (Recidiviz/recidiviz-data#42753)
## Description of the change

Temporarily disables the mental health assessment criteria for AZ admin
supervision until Recidiviz/recidiviz-data#42752 is addressed and we have accurate data in prod.
Currently we see over 300 fewer clients eligible for this opportunity in
prod than in staging (~90% of all clients) because of this missing data.

I've run a sandbox (available at
`recidiviz-123.jkgerig_az_adm_task_eligibility_spans_us_az.transfer_to_administrative_supervision_materialized`)
and see that all the clients eligible in staging are now eligible in
prod with this fix.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b391ba55159b459f59b205fb2dd61cdfd454f448