[US_AZ] Update case manager relationship periods ingest view (Recidiviz/recidiviz-data#42792)
## Description of the change

Quick fix to the client staff relationship periods ingest view in AZ to
account for new raw data we received that shows overlapping periods. A
more thorough investigation is tracked in Recidiviz/recidiviz-data#42789, but this PR simply
updates the `PARTITION` clause we use to determine the next assignment
date.

Sandbox ingest pipeline
[here](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-05-21_12_31_29-9126919981185976100;graphView=0?pageState=(%22dfTime%22:(%22s%22:%222025-05-21T19:31:29.816Z%22,%22e%22:%222025-05-21T19:49:02.881Z%22))&project=recidiviz-staging)

Comparison of view outputs
[here](https://docs.google.com/spreadsheets/d/1adaG8zJGys-_IfRfPR3HhT-MLMFXu5Ggbq86egVMKzk/edit?usp=sharing).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#42759 and Recidiviz/recidiviz-data#42762 but needs to be deployed before those can be
closed.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: cf418897fc9fa30dc738f67b204bd9db326669ac