[Testing] Prune raw data fixtures for non code files (Recidiviz/recidiviz-data#42465)
## Description of the change

This PR updates our fixture generation script to keep the earliest and
latest raw data row for dependencies with a defined set of primary keys.

I updated the MO and TN sentencing tests I was familiar with from Recidiviz/recidiviz-data#42424
and reduced the fixture sizes (and didn't need to update the test
output). I did need to update the TAK026 file to have primary keys. I
think this change is correct, but would like another set of eyes on it.

Other PRs will follow for a solution to code files.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Supports Recidiviz/recidiviz-data#42424

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: fc0eaf74cbc67ad9dea3c14d2e68de12b2dd1571