[US_NE] Add optional qualified table names arg to export script (Recidiviz/recidiviz-data#42850)
## Description of the change

By default, we run the NE SQL server table export on all tables for
which we have file tags. However, we may want to explore data in tables
for which we do not yet have raw file configs. In this case, you can run
the script ad hoc and supply a comma separated list of qualified table
names to export (ex `--qualified-table-names
DCS_WEB.table1,DCS_MVS.table2`), as well as a destination bucket if you
do not want to clutter the ingest bucket, and we will export the table
using the default config for the region and upload it to the destination
bucket. If you want a table to become a regular part of the export
process you need to add a raw file config for it.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

 Recidiviz/recidiviz-data#41296

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 91a3b4c360a3d8807ee64fabf5a09e968f49b0e5